Detect end of code stack while tokenizing function
The function MEMINT_TokenizeFunction had a problem with processing the
very last function in the code stack, because it could not detect its
end. This usually resulted in a crash when using jumps (repeat, etc.),
because MEMINT_TokenizeFunction would run "indefinitely".

Detecting the end of the last function is not trivial, because the code
stack is typically larger then necessary (padded with zero, which is a
valid parser token, i.e. zPAR_OP_PLUS). The solution offered here only
prevents the crash but results in long waiting time. So it's not optimal
and I'd welcome any suggestions.