Remove author prefix for fatal zSpy messages
Messages sent to the zSpy need to have a valid author prefix consisting
of an alphabetic character followed by a colon in order to be processed.
In Ikarus that is "Q:". Messages of type zERR_TYPE_FATAL do not require
any author prefix and will evoke a message box regardless. This change
here skips prefixing fatal messages for reader-friendlier fatal error
boxes.